social-media-growth-ai/requirements.txt
# Social Media Growth AI - Requirements

# Core Dependencies
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0

# API Frameworks
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
httpx>=0.24.0
pytest-mock>=3.10.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.1
isort>=5.12.0

# Documentation
mkdocs>=1.4.2
mkdocs-material>=9.1.5

# Facebook API
facebook-sdk==3.1.0
facebook-business==19.0.0

# Data processing
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.1.post1

# Data Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Export Functionality
openpyxl==3.1.2
XlsxWriter==3.1.9  # Enhanced Excel export capabilities
Pillow==10.2.0    # Required for image handling in Excel exports 